rename shema_loader, add local  and remote  xml validation, add validation tests
